requirements.txt
httpx>=0.27.0
pydantic>=2.0
pyyaml>=6.0
plyer>=2.1.0
flask>=3.0
